backend/requirements-new.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
bitsandbytes>=0.39.0
datasets>=2.12.0
accelerate>=0.20.0

# GPU & Resource Management
GPUtil>=1.4.0
psutil>=5.9.0
pynvml>=11.5.0

# Experiment Tracking
wandb>=0.15.0
mlflow>=2.4.0
tensorboard>=2.13.0

# Security
PyJWT>=2.7.0
bcrypt>=4.0.0
python-multipart>=0.0.6
cryptography>=41.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# File Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0

# Async Task Management
celery>=5.3.0
redis>=4.5.0

# Monitoring
prometheus-client>=0.17.0

# Cloud Storage
boto3>=1.28.0
google-cloud-storage>=2.10.0
azure-storage-blob>=12.17.0

# Existing dependencies
fastapi>=0.110.0
motor>=3.3.0
pydantic>=2.12.0
python-dotenv>=1.2.0
uvicorn>=0.25.0